fix: adicionando o novo header em todas as paginas da aplicacao.
