+++ src/Components/Header/style.js
@@ -1,30 +1,17 @@
 import styled from 'styled-components';
 
-export const Container = styled.header`
-  position: fixed;
-  top:0;
-  left: 0;
+export const Container = styled.div`
   width: 100%;
-  height: 50px;
+  height: 65px;
+  background: #DF8D53;
   display: flex;
-  justify-content: space-between;
   align-items: center;
-  background-color: ${({ theme }) => theme.colors.primary.dark};
-  padding: 0px 10px;
+  justify-content: space-between;
+  padding: 0px 12px;
 
-  img{
-    width: 200px;
-    height: 50px;
-  }
-`;
-
-export const ButtonArea = styled.div`
-  button{
-    width: 100px;
-  }
-
-  button ~ button{
-      margin-left: 12px;
+  .iconMenu{
+    background: none;
+    border: none;
   }
 
   a {
@@ -32,8 +19,46 @@ export const ButtonArea = styled.div`
     color: ${({ theme }) => theme.colors.primary.dark};
   }
 
+  span{
+    cursor: pointer;
+  }
+
+  .navBar{
+    display: flex;
+    gap: 8px;
+
+    button{
+    width: 100px;
+    height: 35px;
+  }
+  }
+
   .delete-button{
-    color: ${({ theme }) => theme.colors.background};
     background: ${({ theme }) => theme.colors.danger.main};
+    color: ${({ theme }) => theme.colors.background};
+    transition: 0.3s;
+  }
+
+  .delete-button:hover{
+    background: ${({ theme }) => theme.colors.danger.dark};
+  }
+
+  @media screen and (min-width: 744px) {
+    padding: 0px 32px;
+
+    .iconMenu{
+      display: none;
+    }
+  }
+
+  @media screen and (max-width: 744px){
+    .navBar{
+      display: none;
+
+      .delete-button{
+       color: ${({ theme }) => theme.colors.background};
+        background: ${({ theme }) => theme.colors.danger.main};
+      }
+    }
   }
 `;
