+++ src/pages/Home/index.js
@@ -1,6 +1,6 @@
 import { ContainerCardsHome } from './style';
 
-// import Header from '../../Components/Header';
+import Header from '../../Components/Header';
 import WelcomeSection from '../../Components/WelcomeSection';
 // import AnimalHome from '../../Components/AnimalHome';
 import Footer from '../../Components/Footer';
@@ -8,12 +8,11 @@ import Dicas from '../../Components/Dicas';
 import MeusAnimaisHome from '../../Components/MeusAnimaisHome';
 import MeusPedidosHome from '../../Components/MeusPedidosHome';
 import AgendamentosHome from '../../Components/AgendamentosHome';
-import NewFooterMenu from '../../Components/NewFooterMenu';
 
 export default function Home() {
   return (
     <>
-      <NewFooterMenu />
+      <Header />
       <WelcomeSection />
       <Dicas />
       <ContainerCardsHome>
